accessible/tests/mochitest/jsat/test_text_navigation_focus.html

<!DOCTYPE html>
<html>
<head>
  <title>Tests AccessFu text entry focus and a11y focus</title>
  <meta charset="utf-8" />
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js">
  </script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js">
  </script>
  <script src="chrome://mochikit/content/chrome-harness.js">
  </script>

  <script type="application/javascript" src="../common.js"></script>
  <script type="application/javascript" src="../browser.js"></script>
  <script type="application/javascript" src="../events.js"></script>
  <script type="application/javascript" src="../role.js"></script>
  <script type="application/javascript" src="../states.js"></script>
  <script type="application/javascript" src="../layout.js"></script>
  <script type="application/javascript" src="jsatcommon.js"></script>

  <script type="application/javascript">
    async function testTextNavigationFocus(doc, runner) {
      let evt;

      evt = await runner.focusSelector("textarea",
        AndroidEvents.VIEW_FOCUSED,
        AndroidEvents.VIEW_ACCESSIBILITY_FOCUSED);
      is(evt[0].editable, true, "focused item is editable");
      is(evt[1].focused, true, "a11y focused item is focused");
      is(evt[1].className, "android.widget.EditText", "editable class");
      runner.eventTextMatches(evt[1],
        ["Please refrain from Mayoneggs during this salmonella scare."]);

      evt = await runner.moveNext("Simple",
        AndroidEvents.VIEW_FOCUSED,
        AndroidEvents.VIEW_ACCESSIBILITY_FOCUSED);
      is(evt[0].editable, false, "focused out of editable");
      runner.eventTextMatches(evt[1], ["So we don't get dessert?", "label"]);
      runner.isFocused("html");

      evt = await runner.moveNext("Simple",
        AndroidEvents.VIEW_ACCESSIBILITY_FOCUSED);
      is(evt.className, "android.widget.EditText", "editable class");
      runner.isFocused("html");

      evt = await runner.activateCurrent(0,
        AndroidEvents.VIEW_CLICKED,
        AndroidEvents.VIEW_FOCUSED,
        AndroidEvents.VIEW_TEXT_SELECTION_CHANGED);
      is(evt[1].editable, true, "focused item is editable");
      is(evt[2].fromIndex, 0, "Cursor at start");
      runner.isFocused("input[type=text]");

      evt = await runner.movePrevious("Simple",
        AndroidEvents.VIEW_FOCUSED,
        AndroidEvents.VIEW_ACCESSIBILITY_FOCUSED);
      is(evt[0].editable, false, "focused out of editable");
      runner.eventTextMatches(evt[1], ["So we don't get dessert?", "label"]);
      runner.isFocused("html");

      evt = await runner.moveNext("Simple",
        AndroidEvents.VIEW_ACCESSIBILITY_FOCUSED);
      is(evt.className, "android.widget.EditText", "editable class");
      runner.isFocused("html");

      // XXX: TEXT_SELECTION_CHANGED should be fired here
      evt = await runner.activateCurrent(0,
        AndroidEvents.VIEW_CLICKED,
        AndroidEvents.VIEW_FOCUSED);
      is(evt[1].editable, true, "focused item is editable");
      runner.isFocused("input[type=text]");

      evt = await runner.blur(AndroidEvents.VIEW_FOCUSED);
      is(evt.editable, false, "Focused out of editable");
    }


    function doTest() {
      var doc = currentTabDocument();

      addA11yLoadEvent(async function() {
        let runner = new AccessFuContentTestRunner();
        await runner.start();

        await testTextNavigationFocus(doc, runner);

        runner.finish();
        closeBrowserWindow();
        SimpleTest.finish();
      }, doc.defaultView);
    }

    SimpleTest.waitForExplicitFinish();
    addLoadEvent(
      function() {
        openBrowserWindow(
          doTest,
          getRootDirectory(window.location.href) + "doc_content_text.html");
        });
  </script>
</head>
<body id="body">

  <a target="_blank"
     title="Add tests for text editing and navigating"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=1459677">Mozilla Bug 1459677</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>
</body>
</html>
